+++ core/impl/src/main/java/com/blazebit/persistence/impl/ResolvingQueryGenerator.java
@@ -597,6 +597,11 @@ public class ResolvingQueryGenerator extends SimpleQueryGenerator {
     }
 
     @Override
+    protected Set<String> getSupportedEnumTypes() {
+        return entityMetamodel.getEnumTypes().keySet();
+    }
+
+    @Override
     protected String getLiteralParameterValue(ParameterExpression expression) {
         Object value = expression.getValue();
         if (value == null) {
@@ -604,7 +609,7 @@ public class ResolvingQueryGenerator extends SimpleQueryGenerator {
         }
 
         if (value != null) {
-            final TypeConverter<Object> converter = (TypeConverter<Object>) TypeUtils.getConverter(value.getClass());
+            final TypeConverter<Object> converter = (TypeConverter<Object>) TypeUtils.getConverter(value.getClass(), getSupportedEnumTypes());
             // We can't render enum values as literals directly, only in the context of a predicate, so we need the BooleanLiteralRenderingContext.PLAIN
             if (converter != null && (!(value instanceof Enum<?>) || getBooleanLiteralRenderingContext() == BooleanLiteralRenderingContext.PLAIN)) {
                 return converter.toString(value);