+++ integration/datanucleus/src/main/java/com/blazebit/persistence/integration/datanucleus/DataNucleusJpaProvider.java
@@ -592,6 +592,11 @@ public class DataNucleusJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean supportsEnumLiteral(ManagedType<?> ownerType, String attributeName, boolean key) {
+        return true;
+    }
+
+    @Override
     public Object getIdentifier(Object entity) {
         return persistenceUnitUtil.getIdentifier(entity);
     }