+++ CHANGELOG.md
@@ -12,7 +12,8 @@ None
 
 ### Bug fixes
 
-None
+* Missing flushes for attributes that were changed in pre-update entity view listeners
+* Render enum literal as parameter if compared with attribute having Hibernate non-`EnhancedUserType`
 
 ### Backwards-incompatible changes